# api/openapi.yaml
openapi: 3.0.3
info:
  title: BagiCore
  description: Go-based comic hosting catalog backend.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
...
servers:
  - url: /api/v0
paths:
...
  /languages:
    get:
      tags:
        - Language
      summary: List language.
      operationId: listLanguage
      parameters:
        - name: page
          in: query
          description: Page number of results.
          schema:
            type: integer
        - name: limit
          in: query
          description: Maximum number of results.
          schema:
            type: integer
        - name: order_by
          in: query
          description: Sort results returned.
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Language list.
          headers:
            X-Total-Count:
              schema:
                type: integer
              description: The total count of language with current filter.
            X-Pagination-Limit:
              schema:
                type: integer
              description: The last page number of language with current filter and limit.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Language'
        default:
          $ref: '#/components/responses/Default'
    post:
      tags:
        - Language
      summary: Add language.
      operationId: addLanguage
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewLanguage'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NewLanguage'
        required: true
      responses:
        '201':
          description: Language added.
          headers:
            Location:
              schema:
                type: string
              description: The path of new language.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Language'
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
  /languages/{ietf}:
    get:
      tags:
        - Language
      summary: Get language.
      operationId: getLanguage
      parameters:
        - name: ietf
          in: path
          description: IETF code of language to return.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Language gets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Language'
        default:
          $ref: '#/components/responses/Default'
    patch:
      tags:
        - Language
      summary: Update language.
      operationId: updateLanguage
      parameters:
        - name: ietf
          in: path
          description: IETF code of language to update.
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetLanguage'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SetLanguage'
        required: true
      responses:
        '200':
          description: Language updated.
          headers:
            Location:
              description: The path of updated language.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Language'
        '204':
          description: Language unmodified.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
    delete:
      tags:
        - Language
      summary: Delete language.
      operationId: deleteLanguage
      parameters:
        - name: ietf
          in: path
          description: IETF code of language to delete.
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Language deleted.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
...
components:
  schemas:
    Object:
      type: object
      properties:
        id:
          type: integer
          format: int64
          x-go-type: uint
          x-go-name: ID
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - createdAt
...
    Language:
      type: object
      allOf:
        - $ref: '#/components/schemas/Object'
        - type: object
          properties:
            ietf:
              type: string
              x-go-name: IETF
            name:
              type: string
          required: 
            - ietf
            - name
    NewLanguage:
      type: object
      properties:
        ietf:
          type: string
          x-go-name: IETF
        name:
          type: string
      required: 
        - ietf
        - name
    SetLanguage:
      type: object
      properties:
        ietf:
          type: string
          nullable: true
          x-go-name: IETF
        name:
          type: string
          nullable: true
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties: 
            status:
              type: string
            message:
              type: string
          required:
            - status
            - message
      required:
        - error
  responses:
    Default:
      description: Unexpected error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT